requirements.txt
python-binance>=1.0.17
pandas>=2.1.0
numpy>=1.24.0
python-dotenv>=1.0.0
matplotlib>=3.7.0
requests>=2.31.0
ta>=0.11.0
